Update README.md
Added a description explaining why NFT images are not visible and how to fix the issue. This is important because users may be confused if they can't see the images. The reason why the images are not visible is that the collection names in the project have been changed, so the NFT images cannot be downloaded using the old names. To fix this issue, users need to update the collection names in the project with the new names. For example, in the digital art section of the gallery, the "Collection Name" value used to be "fake-mutant-ape-yacht-club", but since the collection has been removed, it has been changed to "boredapeyachtclubgoerli". You can use the details panel of the actor in the digital art section for changing values. Once the collection names have been updated, the NFT images will be visible.